PRLC100-346  Sonar Duplicate Block Fix For Child Matters
