+++ src/main/steps/c100-rebuild/child-details/childDetailPostController.test.ts
@@ -6,7 +6,7 @@ import { CommonContent } from '../../common/common.content';
 import ChildDetailsPostController from './childDetailPostController';
 import { generateContent } from './personal-details/content';
 
-describe('ChildDetailsPostController Post Controller', () => {
+describe('ChildDetailsPostController - personal detail common Post Controller', () => {
   const commonContent = {
     language: 'en',
     userCase: {
@@ -87,6 +87,222 @@ describe('ChildDetailsPostController Post Controller', () => {
           type: 'hidden',
           labelHidden: true,
           value: 'pd',
+        },
+      },
+    } as unknown as FormContent;
+    const controller = new ChildDetailsPostController(mockFormContent.fields);
+    const language = 'en';
+    const req = mockRequest({
+      params: {
+        childId: '7483640e-0817-4ddc-b709-6723f7925474',
+      },
+      body: {
+        saveAndComeLater: true,
+      },
+      session: {
+        lang: language,
+        userCase: {
+          ...commonContent.userCase,
+        },
+      },
+    });
+    const res = mockResponse();
+    generateContent(commonContent);
+    await controller.post(req, res);
+
+    expect(res.redirect).toHaveBeenCalled();
+  });
+});
+
+describe('ChildDetailsPostController - parential responsibility common Post Controller', () => {
+  const commonContent = {
+    language: 'en',
+    userCase: {
+      cd_children: [
+        {
+          id: '7483640e-0817-4ddc-b709-6723f7925474',
+          firstName: 'Bob',
+          lastName: 'Silly',
+          personalDetails: {
+            dateOfBirth: {
+              year: '',
+              month: '',
+              day: '',
+            },
+            isDateOfBirthUnknown: 'Yes',
+            approxDateOfBirth: {
+              year: '1987',
+              month: '12',
+              day: '12',
+            },
+            sex: 'Male',
+          },
+          childMatters: {
+            needsResolution: [],
+          },
+          parentialResponsibility: {
+            statement: 'fgfdgfg',
+          },
+        },
+      ],
+    },
+    additionalData: {
+      req: {
+        params: {
+          childId: '7483640e-0817-4ddc-b709-6723f7925474',
+        },
+      },
+    },
+  } as unknown as CommonContent;
+
+  test('Should navigagte to the next page when there are no errors when continue button is clicked', async () => {
+    const mockFormContent = {
+      fields: {
+        _ctx: {
+          type: 'hidden',
+          labelHidden: true,
+          value: 'pr',
+        },
+      },
+    } as unknown as FormContent;
+    const controller = new ChildDetailsPostController(mockFormContent.fields);
+    const language = 'en';
+    const req = mockRequest({
+      params: {
+        childId: '7483640e-0817-4ddc-b709-6723f7925474',
+      },
+      body: {
+        onlycontinue: true,
+      },
+      session: {
+        lang: language,
+        userCase: {
+          ...commonContent.userCase,
+        },
+      },
+    });
+    const res = mockResponse();
+    generateContent(commonContent);
+    await controller.post(req, res);
+
+    expect(res.redirect).toHaveBeenCalled();
+  });
+
+  test('Should update case when save and come back button is clicked', async () => {
+    const mockFormContent = {
+      fields: {
+        _ctx: {
+          type: 'hidden',
+          labelHidden: true,
+          value: 'pr',
+        },
+      },
+    } as unknown as FormContent;
+    const controller = new ChildDetailsPostController(mockFormContent.fields);
+    const language = 'en';
+    const req = mockRequest({
+      params: {
+        childId: '7483640e-0817-4ddc-b709-6723f7925474',
+      },
+      body: {
+        saveAndComeLater: true,
+      },
+      session: {
+        lang: language,
+        userCase: {
+          ...commonContent.userCase,
+        },
+      },
+    });
+    const res = mockResponse();
+    generateContent(commonContent);
+    await controller.post(req, res);
+
+    expect(res.redirect).toHaveBeenCalled();
+  });
+});
+
+describe('ChildDetailsPostController -  child matter common Post Controller', () => {
+  const commonContent = {
+    language: 'en',
+    userCase: {
+      cd_children: [
+        {
+          id: '7483640e-0817-4ddc-b709-6723f7925474',
+          firstName: 'Bob',
+          lastName: 'Silly',
+          personalDetails: {
+            dateOfBirth: {
+              year: '',
+              month: '',
+              day: '',
+            },
+            isDateOfBirthUnknown: 'Yes',
+            approxDateOfBirth: {
+              year: '1987',
+              month: '12',
+              day: '12',
+            },
+            sex: 'Male',
+          },
+          childMatters: {
+            needsResolution: [],
+          },
+          parentialResponsibility: {
+            statement: 'fgfdgfg',
+          },
+        },
+      ],
+    },
+    additionalData: {
+      req: {
+        params: {
+          childId: '7483640e-0817-4ddc-b709-6723f7925474',
+        },
+      },
+    },
+  } as unknown as CommonContent;
+
+  test('Should navigagte to the next page when there are no errors when continue button is clicked', async () => {
+    const mockFormContent = {
+      fields: {
+        _ctx: {
+          type: 'hidden',
+          labelHidden: true,
+          value: 'cm',
+        },
+      },
+    } as unknown as FormContent;
+    const controller = new ChildDetailsPostController(mockFormContent.fields);
+    const language = 'en';
+    const req = mockRequest({
+      params: {
+        childId: '7483640e-0817-4ddc-b709-6723f7925474',
+      },
+      body: {
+        onlycontinue: true,
+      },
+      session: {
+        lang: language,
+        userCase: {
+          ...commonContent.userCase,
+        },
+      },
+    });
+    const res = mockResponse();
+    generateContent(commonContent);
+    await controller.post(req, res);
+
+    expect(res.redirect).toHaveBeenCalled();
+  });
+
+  test('Should update case when save and come back button is clicked', async () => {
+    const mockFormContent = {
+      fields: {
+        _ctx: {
+          type: 'hidden',
+          labelHidden: true,
+          value: 'cm',
         },
       },
     } as unknown as FormContent;
